move CharsetEncoding class to Core
